brand-monitoring/brand_monitoring_flow/pyproject.toml
[project]
name = "brand_monitoring_flow"
version = "0.1.0"
description = "brand-monitoring-flow using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.102.0,<1.0.0",
    "langchain-groq>=0.1.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
]

[project.scripts]
kickoff = "brand_monitoring_flow.main:kickoff"
plot = "brand_monitoring_flow.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
